Fix warnings

+++ DailyCodingProblem/Problem_40.h
@@ -43,7 +43,7 @@ int nonRepeatInt( vector<int> v )
   int result = 0;
   for( int i = 0; i < 32; i++ )
   {
-    if( result_arr[i] ) { result += pow( 2, i ); }
+    if( result_arr[i] ) { result += (int)pow( 2, i ); }
   }
 
   return result;
